requirements.txt
Flask~=3.1.1
SQLAlchemy
python-dotenv~=1.1.1
psycopg2-binary
protobuf~=5.29.5
Flask-Migrate~=4.1.0
requests
aiohttp